refactor(instagram): better code organization
