+++ integrations/instagram/src/index.ts
@@ -1,16 +1,11 @@
-import { IntegrationContext } from '@botpress/sdk'
 import { sentry as sentryHelpers } from '@botpress/sdk-addons'
-import { MessengerClient, MessengerTypes } from 'messaging-api-messenger'
 import queryString from 'query-string'
 import { idTag } from './const'
+import { handleMessage } from './misc/incoming-message'
+import { sendMessage } from './misc/outgoing-message'
+import { InstagramPayload } from './misc/types'
+import { formatGoogleMapLink, getCarouselMessage, getChoiceMessage, getMessengerClient } from './misc/utils'
 import * as bp from '.botpress'
-
-type Channels = bp.Integration['channels']
-type Messages = Channels[keyof Channels]['messages']
-type MessageHandler = Messages[keyof Messages]
-type MessageHandlerProps = Parameters<MessageHandler>[0]
-type IntegrationLogger = Parameters<bp.IntegrationProps['handler']>[0]['logger']
-type InstagramUserProfile = MessengerTypes.User & { username: string }
 
 const integration = new bp.Integration({
   register: async () => {},
@@ -191,218 +186,3 @@ export default sentryHelpers.wrapIntegration(integration, {
   environment: bp.secrets.SENTRY_ENVIRONMENT,
   release: bp.secrets.SENTRY_RELEASE,
 })
-
-type Carousel = bp.channels.channel.carousel.Carousel
-type Card = bp.channels.channel.card.Card
-type Choice = bp.channels.channel.choice.Choice
-type Dropdown = bp.channels.channel.dropdown.Dropdown
-type Location = bp.channels.channel.location.Location
-
-type InstagramAttachment = InstagramPostbackAttachment | InstagramSayAttachment | InstagramUrlAttachment
-
-type InstagramPostbackAttachment = {
-  type: 'postback'
-  title: string
-  payload: string
-}
-
-type InstagramSayAttachment = {
-  type: 'postback'
-  title: string
-  payload: string
-}
-
-type InstagramUrlAttachment = {
-  type: 'web_url'
-  title: string
-  url: string
-}
-
-function formatCardElement(payload: Card) {
-  const buttons: InstagramAttachment[] = []
-
-  payload.actions.forEach((action) => {
-    switch (action.action) {
-      case 'postback':
-        buttons.push({
-          type: 'postback',
-          title: action.label,
-          payload: `postback:${action.value}`,
-        })
-        break
-      case 'say':
-        buttons.push({
-          type: 'postback',
-          title: action.label,
-          payload: `say:${action.value}`,
-        })
-        break
-      case 'url':
-        buttons.push({
-          type: 'web_url',
-          title: action.label,
-          url: action.value,
-        })
-        break
-      default:
-        break
-    }
-  })
-  return {
-    title: payload.title,
-    image_url: payload.imageUrl,
-    subtitle: payload.subtitle,
-    buttons,
-  }
-}
-
-function getMessengerClient(configuration: bp.configuration.Configuration) {
-  return new MessengerClient({
-    accessToken: configuration.accessToken,
-  })
-}
-
-function getCarouselMessage(payload: Carousel): MessengerTypes.AttachmentMessage {
-  return {
-    attachment: {
-      type: 'template',
-      payload: {
-        templateType: 'generic',
-        elements: payload.items.map(formatCardElement),
-      },
-    },
-  }
-}
-
-function getChoiceMessage(payload: Choice | Dropdown): MessengerTypes.AttachmentMessage {
-  return {
-    attachment: {
-      type: 'template',
-      payload: {
-        templateType: 'generic',
-      },
-    },
-    quickReplies: payload.options.map((choice) => ({
-      contentType: 'text',
-      title: choice.label,
-      payload: choice.value,
-    })),
-  }
-}
-
-type SendMessageProps = Pick<MessageHandlerProps, 'ctx' | 'conversation' | 'ack'>
-
-async function sendMessage(
-  { ack, ctx, conversation }: SendMessageProps,
-  send: (client: MessengerClient, recipientId: string) => Promise<{ messageId: string }>
-) {
-  const messengerClient = getMessengerClient(ctx.configuration)
-  const recipientId = getRecipientId(conversation)
-  await send(messengerClient, recipientId)
-  await ack({
-    tags: {
-      // TODO: declare in definition
-      // [idTag]: messageId,
-    },
-  })
-}
-
-export function getRecipientId(conversation: SendMessageProps['conversation']): string {
-  const recipientId = conversation.tags[idTag]
-
-  if (!recipientId) {
-    throw Error(`No recipient id found for user ${conversation.id}`)
-  }
-
-  return recipientId
-}
-
-async function handleMessage(
-  message: InstagramMessage,
-  {
-    client,
-    ctx,
-    logger,
-  }: { client: bp.Client; ctx: IntegrationContext<bp.configuration.Configuration>; logger: IntegrationLogger }
-) {
-  if (message?.message?.text) {
-    logger.forBot().debug('Received text message from Instagram:', message.message.text)
-    const { conversation } = await client.getOrCreateConversation({
-      channel: 'channel',
-      tags: {
-        [idTag]: message.sender.id,
-      },
-    })
-
-    const { user } = await client.getOrCreateUser({
-      tags: {
-        [idTag]: message.sender.id,
-      },
-    })
-
-    if (!user.pictureUrl || !user.name) {
-      try {
-        const messengerClient = getMessengerClient(ctx.configuration)
-        const userProfile = (await messengerClient.getUserProfile(message.sender.id, {
-          // username is an available field for instagram ids -> https://developers.facebook.com/docs/instagram-basic-display-api/guides/getting-profiles-and-media
-          fields: ['id', 'name', 'profile_pic', 'username'] as any,
-        })) as InstagramUserProfile
-
-        logger.forBot().debug('Fetched latest Instagram user profile: ', userProfile)
-
-        const fieldsToUpdate = {
-          pictureUrl: userProfile?.profilePic,
-          name: userProfile?.name || userProfile?.username,
-        }
-        if (fieldsToUpdate.pictureUrl || fieldsToUpdate.name) {
-          await client.updateUser({ ...user, ...fieldsToUpdate })
-        }
-      } catch (error) {
-        logger.forBot().error('Error while fetching user profile from Instagram', error)
-      }
-    }
-
-    await client.createMessage({
-      type: 'text',
-      tags: {
-        // TODO: declare in definition
-        // [idTag]: message.message.mid
-      },
-      userId: user.id,
-      conversationId: conversation.id,
-      payload: { text: message.message.text },
-    })
-  }
-}
-
-function formatGoogleMapLink(payload: Location) {
-  return `https://www.google.com/maps/search/?api=1&query=${payload.latitude},${payload.longitude}`
-}
-
-type InstagramPayload = {
-  object: string
-  entry: InstagramEntry[]
-}
-
-type InstagramEntry = {
-  id: string
-  time: number
-  messaging: InstagramMessage[]
-}
-
-type InstagramMessage = {
-  sender: { id: string }
-  recipient: { id: string }
-  timestamp: number
-  message?: {
-    mid: string
-    text: string
-    quick_reply?: { payload: string }
-    attachments?: { type: string; payload: { url: string } }[]
-  }
-  postback?: {
-    mid: string
-    payload: string
-    title: string
-  }
-}
